+++ src/Exporter/ExporterInterface.php
@@ -12,7 +12,9 @@
 namespace HeimrichHannot\ContaoExporterBundle\Exporter;
 
 
+use HeimrichHannot\ContaoExporterBundle\Exception\ExporterConfigurationException;
 use HeimrichHannot\ContaoExporterBundle\Exception\ExportNotPossibleException;
+use HeimrichHannot\ContaoExporterBundle\Exception\ExportTypeNotSupportedException;
 use HeimrichHannot\ContaoExporterBundle\Model\ExporterModel;
 
 interface ExporterInterface
@@ -33,7 +35,10 @@ interface ExporterInterface
      * @param null $entity
      * @param array $fields
      * @return bool
+     *
      * @throws ExportNotPossibleException
+     * @throws ExportTypeNotSupportedException
+     * @throws ExporterConfigurationException
      */
     public function export(ExporterModel $config = null, $entity = null, array $fields = []): bool;
 }