+++ src/Resources/contao/languages/de/default.php
@@ -14,9 +14,3 @@ $arrLang['export_xls'] = ['Exportieren (Excel)', 'Als Excel-Datei exportieren'];
  */
 $arrLang['exporter']['unformatted'] = ' (unformatiert)';
 $arrLang['exporter']['export'] = 'Exportieren';
-
-/**
- * Messages
- */
-$arrLang['exporter']['noConfigFound'] = 'Für die gewählte Aktion wurde keine Exporter-Konfiguration gefunden. Legen Sie diese über den Menüpunkt "Exporter" im Backend an.';
-$arrLang['exporter']['exporterNotPossible'] = 'Der Exportvorgang konnte nicht durchgeführt werden. Bitte wenden Sie sich an den Administrator der Website.';