+++ src/Resources/contao/languages/en/default.php
@@ -15,9 +15,3 @@ $arrLang['export_xls'] = ['Export (Excel)', 'Export as Excel file'];
 
 $arrLang['exporter']['unformatted'] = ' (unformatted)';
 $arrLang['exporter']['export'] = 'Export';
-
-/**
- * Messages
- */
-$arrLang['exporter']['noConfigFound'] = 'No exporter config could be found for the chosen action. Please create one in the backend module "Exporter" on the left.';
-$arrLang['exporter']['exporterNotPossible'] = 'The export process could not be performed. Please contact the administrator of the website.';